Added  date and time knowledge + token limit check
Glitch now knows what the current date and time are. He *should* be able to use that info and infer what the date and time in other timezones are, but in practice he's pretty bad at it.

He also lets users know if he ran out of tokens to generate his message.

Added some variables to the top of the script that I can use to tweak certain parts of him

Updated his personality